pyproject.toml
[project]
name = "vae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pytz>=2025.2",
    "tqdm>=4.67.3",
    "wandb>=0.25.0",
    "torch==2.9.1",
    "torchvision>=0.24.1",
    "dotenv>=0.9.9",
    "hydra-core>=1.3.2",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu129", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu129", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
